% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhoc.twosigma.R
\name{adhoc.twosigma}
\alias{adhoc.twosigma}
\title{adhoc.twosigma: Perform the ad hoc method described in TWO-SIGMA paper}
\usage{
adhoc.twosigma(count, mean_covar, zi_covar, id, weights = rep(1,
  length(count)), control = glmmTMBControl(optCtrl = list(iter.max =
  1e+05, eval.max = 1e+05, step.max = 1e-05, step.min = 1e-05, rel.tol =
  1e-05, x.tol = 1e-05)))
}
\arguments{
\item{count}{Vector of non-negative integer read counts.}

\item{mean_covar}{Covariates for the (conditional) mean model. Must be a matrix (without an intercept column) or = 1 to indicate an intercept only model.}

\item{zi_covar}{Covariates for the zero-inflation model. Must be a matrix (without an intercept column), = 1 to indicate an intercept only model, or = 0 to indicate no zero-inflation model desired.}

\item{id}{Vector of individual-level ID's. Used as predictor in ANOVA model.
@param weights weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed. Passed into zeroinfl function
@return P-value from the ANOVA F test.}
}
\description{
adhoc.twosigma: Perform the ad hoc method described in TWO-SIGMA paper
}
