% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.twosigma.R
\name{lr.twosigma}
\alias{lr.twosigma}
\title{Conveinent wrapper function for performing joint likelihood ratio tests using the TWO-SIGMA model of ...}
\usage{
lr.twosigma(count, mean_covar, zi_covar, contrast, mean_re = FALSE,
  zi_re = FALSE, id, adhoc = FALSE, adhoc_thresh = 0.1,
  disp_covar = NULL, weights = rep(1, length(count)),
  control = glmmTMBControl())
}
\arguments{
\item{count}{Vector of non-negative integer read counts.}

\item{mean_covar}{Covariates for the (conditional) mean model. Must be a matrix (without an intercept column) or a vector if a single covariate is being tested.}

\item{zi_covar}{Covariates for the zero-inflation model. Must be a matrix (without an intercept column) or a vector if a single covariate is being tested.}

\item{contrast}{Either a string indicating the column name of the covariate to test or an integer referring to its column position in BOTH the mean_covar and zi_covar matrices (if the two matrices differ using a string name is preferred). Argument is ignored if mean_covar and zi_covar are both a single covariate (that covariate is assumed of interest).}

\item{mean_re}{Should random intercepts be included in the (conditional) mean model?}

\item{zi_re}{Should random intercepts be included in the zero-inflation model?}

\item{id}{Vector of individual-level ID's. Used for random effect prediction and the adhoc method but required regardless.}

\item{adhoc}{Should the adhoc method be used by default to judge if random effects are needed?}

\item{adhoc_thresh}{Value below which the adhoc p-value is deemed significant (and thus RE are deemed necessary). Only used if adhoc==TRUE.}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix or = 1 to indicate an intercept only model.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in glmmTMB.}
}
\value{
A list containing glmmTMB objects of model fits under the null and alternative, the 2 d.f. Likelihood Ratio statistic, and the p-value.
}
\description{
Conveinent wrapper function for performing joint likelihood ratio tests using the TWO-SIGMA model of ...
}
\section{Details}{

This function assumes that the variable being tested is in both components of the model (and thus that the zero-inflation component exists and contains more than an Intercept). Users wishing to do fixed effect testing in other cases or specify custom model formulas they will need to construct the statistics themselves using either two separate calls to \code{twosigma} or the \code{lr.twosigma_custom} fumction. If adhoc==TRUE, any input in mean_re and zi_re will be ignored. If either model fails to converge, or the LR statistic is negative, both the statistic and p-value are assigned as NA.
}

