% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigma.R
\name{twosigma}
\alias{twosigma}
\title{TWO-SIGMA: Fit the TWO-component SInGle cell Model-based Association method of...}
\usage{
twosigma(count, mean_covar = NULL, zi_covar = NULL, mean_re = NULL,
  zi_re = NULL, id, mean_form = NULL, zi_form = NULL,
  disp_covar = NULL, weights = rep(1, length(count)),
  control = glmmTMBControl())
}
\arguments{
\item{count}{Vector of non-negative integer read counts. If specifying custom formula(s) via the arguments mean_form and zi_form the expression in mean_form will supersede.}

\item{mean_covar}{Covariates for the (conditional) mean model. Must be a matrix (without an intercept column) or = 1 to indicate an intercept only model.}

\item{zi_covar}{Covariates for the zero-inflation model. Must be a matrix (without an intercept column), = 1 to indicate an intercept only model, or = 0 to indicate no zero-inflation model desired.}

\item{mean_re}{Should random intercepts be included in the (conditional) mean model?}

\item{zi_re}{Should random intercepts be included in the zero-inflation model?}

\item{id}{Vector of individual-level ID's. Used for random effect prediction.}

\item{mean_form}{Custom two-sided model formula for the (conditional) mean model. Only necessary if it is desired to include more complex random effects beyond a random intercept. Formula is passed directly into glmmTMB with random effects specified as in the lme4 package. If specified takes precedence over the \code{mean_re} and \code{mean_covar} arguments. Users should ensure that the dependent variable matches the argument to the parameter "count."}

\item{zi_form}{Custom one-sided model formula for the zero-inflation model. Only necessary if it is desired to include more complex random effects beyond a random intercept. Formula is passed directly into glmmTMB with random effects specified as in lme4. If specified takes precedence over the \code{zi_re} and \code{zi_covar} arguments.}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}.}
}
\value{
An object of class \code{glmmTMB}. See Details for a note about the estimated overdispersion parameter.
}
\description{
TWO-SIGMA: Fit the TWO-component SInGle cell Model-based Association method of...
}
\section{Details}{
 Note that the overdispersion estimate returned is the inverse of phi in the parameterization used in TWO-SIGMA.  See ?nbinom2
}

