% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigmag.R
\name{twosigmag}
\alias{twosigmag}
\title{Gene set testing for single-cell RNA-sequencing data adjusting for inter-gene correlation.}
\usage{
twosigmag(count_matrix, index_test, index_ref = NULL,
  all_as_ref = FALSE, covar_to_test, mean_covar, zi_covar,
  mean_re = TRUE, zi_re = TRUE, id, rho = NULL,
  allow_neg_corr = FALSE, adhoc = FALSE, adhoc_thresh = 0.1,
  disp_covar = NULL, return_summary_fits = FALSE, weights = rep(1,
  ncol(count_matrix)), control = glmmTMBControl())
}
\arguments{
\item{count_matrix}{Matrix of non-negative integer read counts. If specifying custom formula(s) via the arguments mean_form and zi_form the expression in mean_form will supersede. It is recommended to make the rownames the gene names for better output.}

\item{index_test}{List of indices corresponding to rows of the count matrix that are in the test set.}

\item{index_ref}{List of indices corresponding to rows of the count matrix that are in the reference set.  If NULL, a reference set is randomly selected of the same size as the test size using genes not in the test set (if all_as_ref=FALSE) or using all other genes (if all_as_ref=TRUE). See \code{all_as_ref}. Must be either NULL or a list with the same length as index_test.}

\item{all_as_ref}{Should all genes not in the test set be used as the reference? If FALSE, a random subset is taken of size equal to the test size.}

\item{covar_to_test}{Covariate used for reporting direction (as Up or Down) of the test set. Either a string indicating the name of the covariate to use or an integer giving its associated position in the mean_covar argument.}

\item{mean_covar}{Covariates for the (conditional) mean model. Must be a matrix (without an intercept column) or = 1 to indicate an intercept only model.}

\item{zi_covar}{Covariates for the zero-inflation model. Must be a matrix (without an intercept column), = 1 to indicate an intercept only model, or = 0 to indicate no zero-inflation model desired.}

\item{mean_re}{Should random intercepts be included in the (conditional) mean model? Ignored if adhoc=TRUE.}

\item{zi_re}{Should random intercepts be included in the zero-inflation model? Ignored if adhoc=TRUE.}

\item{id}{Vector of individual-level (sample-level) ID's. Used to estimate inter-gene correlation and for random effect prediction (if present).}

\item{rho}{Inter-gene correlation value. If NULL (default), estimated using model residuals from TWO-SIGMAG.}

\item{allow_neg_corr}{Should negative correlation values be allowed? If FALSE, correlation is set to zero (leads to conservative inference).}

\item{adhoc}{Should the adhoc method be used by default to judge if random effects are needed? Defaults to FALSE. Setting to TRUE means that different genes could have different models used, which may not be desirable for interpretability.}

\item{adhoc_thresh}{Value below which the adhoc p-value is deemed significant (and thus RE are deemed necessary). Only used if adhoc==TRUE.}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model. Defaults to NULL for constant dispersion.}

\item{return_summary_fits}{If TRUE, returns a summary.glmmTMB object for the Alternative model.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}. See \code{?glmmTMBControl}.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{gene_summary_fits: }}{summary.glmmTMB objects for each gene from the alternative model (if return_summary_fits=TRUE)}
\item{\code{LR_stats_gene_level_all: }}{Gives all gene-level likelihood ratio statistics.  Order matches the order of the inputted count matrix.}
\item{\code{p.vals_gene_level: }}{Gives p-values associated with \code{LR_stats_gene_level_all}.}
\item{\code{set_p.val: }}{Vector of unadjusted set-level p-values. Order matches the order of inputted test sets.}
\item{\code{set_p.val_ttest: }}{Vector of unadjusted set-level p-values using the t-test. Order matches the order of inputted test sets.}
\item{\code{avg_logFC_gene_level: }}{Gives the average logFC for the covariate specified in the mean_covar_logFC argument.}
\item{\code{direction: }}{Reports whether the test set tends to be Up or Down Regulated based on the covariate specified in the mean_covar_logFC argument.}
\item{\code{corr: }}{Vector of estimated inter-gene correlations for each test set. Order matches the order of inputted test sets.}
\item{\code{test_sets: }}{Vector of numeric indices corresponding to genes in each test set.}
\item{\code{ref_sets: }}{Vector of numeric indices corresponding to the genes in each reference set.}
}
}
\description{
Gene set testing for single-cell RNA-sequencing data adjusting for inter-gene correlation.
}
\section{Details}{
  If adhoc=TRUE, any input in mean_re and zi_re will be ignored.
}

