% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigmag_custom.R
\name{twosigmag_custom}
\alias{twosigmag_custom}
\title{TWO-SIGMA: Fit the TWO-component SInGle cell Model-based Association method of...}
\usage{
twosigmag_custom(count_matrix, index, mean_form_alt, zi_form_alt,
  mean_form_null, zi_form_null, id, rho = NULL, disp_covar = NULL,
  verbose_output = FALSE, weights = rep(1, length(count_matrix[1, ])),
  control = glmmTMBControl())
}
\arguments{
\item{count_matrix}{Matrix of non-negative integer read counts. If specifying custom formula(s) via the arguments mean_form and zi_form the expression in mean_form will supersede.}

\item{id}{Vector of individual-level ID's. Used for random effect prediction and the adhoc method but required regardless.}

\item{rho}{Inter-gene correlation value. If NULL (default), estimated using TWO-SIGMA model residuals.}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}.}

\item{mean_re}{Should random intercepts be included in the (conditional) mean model? Ignored if adhoc=TRUE.}

\item{zi_re}{Should random intercepts be included in the zero-inflation model? Ignored if adhoc=TRUE.}
}
\value{
An object of class \code{glmmTMB}.
}
\description{
TWO-SIGMA: Fit the TWO-component SInGle cell Model-based Association method of...
}
\section{Details}{
  If adhoc=TRUE, any input in mean_re and zi_re will be ignored.
}

