% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigmag_list.R
\name{twosigmag_list}
\alias{twosigmag_list}
\title{Gene set testing adjusting for inter-gene correlation using the TWO-SIGMA model of ...}
\usage{
twosigmag_list(count_matrix, index_test, index_ref = NULL, contrast,
  mean_covar, zi_covar, mean_re = TRUE, zi_re = TRUE, id, rho = NULL,
  adhoc = TRUE, adhoc_thresh = 0.1, disp_covar = NULL,
  return_fits = FALSE, weights = rep(1, length(count_matrix[1, ])),
  control = glmmTMBControl())
}
\arguments{
\item{count_matrix}{Matrix of non-negative integer read counts. If specifying custom formula(s) via the arguments mean_form and zi_form the expression in mean_form will supersede.}

\item{mean_covar}{Covariates for the (conditional) mean model. Must be a matrix (without an intercept column) or = 1 to indicate an intercept only model.}

\item{zi_covar}{Covariates for the zero-inflation model. Must be a matrix (without an intercept column), = 1 to indicate an intercept only model, or = 0 to indicate no zero-inflation model desired.}

\item{mean_re}{Should random intercepts be included in the (conditional) mean model? Ignored if adhoc=TRUE.}

\item{zi_re}{Should random intercepts be included in the zero-inflation model? Ignored if adhoc=TRUE.}

\item{id}{Vector of individual-level ID's. Used for random effect prediction and the adhoc method but required regardless.}

\item{rho}{Inter-gene correlation value. If NULL (default), estimated using model residuals from TWO-SIGMAG.}

\item{adhoc}{Should the adhoc method be used by default to judge if random effects are needed? Defaults to FALSE. Setting to TRUE means that different genes could have different models used, which may not be desirable for interpretability.}

\item{adhoc_thresh}{Value below which the adhoc p-value is deemed significant (and thus RE are deemed necessary). Only used if adhoc==TRUE.}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model. Defaults to NULL for constant dispersion.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}.}
}
\value{
An object of class \code{glmmTMB}.
}
\description{
Gene set testing adjusting for inter-gene correlation using the TWO-SIGMA model of ...
}
\section{Details}{
  If adhoc=TRUE, any input in mean_re and zi_re will be ignored.
}

