% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigma.R
\name{twosigma}
\alias{twosigma}
\title{TWO-SIGMA: Fit the TWO-component SInGle cell Model-based Association method of...}
\usage{
twosigma(count, mean_covar, zi_covar, mean_re = TRUE, zi_re = TRUE, id,
  adhoc = TRUE, adhoc_thresh = 0.1, disp_covar = NULL,
  weights = rep(1, length(count)), control = glmmTMBControl())
}
\arguments{
\item{count}{Vector of non-negative integer read counts. If specifying custom formula(s) via the arguments mean_form and zi_form the expression in mean_form will supersede.}

\item{mean_covar}{Covariates for the (conditional) mean model. Must be a matrix (without an intercept column) or = 1 to indicate an intercept only model.}

\item{zi_covar}{Covariates for the zero-inflation model. Must be a matrix (without an intercept column), = 1 to indicate an intercept only model, or = 0 to indicate no zero-inflation model desired.}

\item{mean_re}{Should random intercepts be included in the (conditional) mean model?}

\item{zi_re}{Should random intercepts be included in the zero-inflation model?}

\item{id}{Vector of individual-level ID's. Used for random effect prediction and the adhoc method but required regardless.}

\item{adhoc}{Should the adhoc method be used by default to judge if random effects are needed?}

\item{adhoc_thresh}{Value below which the adhoc p-value is deemed significant (and thus RE are deemed necessary). Only used if adhoc==TRUE.}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}.}
}
\value{
An object of class \code{glmmTMB}.
}
\description{
TWO-SIGMA: Fit the TWO-component SInGle cell Model-based Association method of...
}
\section{Details}{
  If adhoc==TRUE, any input in mean_re and zi_re will be ignored.  See ?nbinom2
}

