% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigma_custom.R
\name{twosigma_custom}
\alias{twosigma_custom}
\title{Fit the TWO-component SInGle cell Model-based Association method of ... with custom user-specified model formulas.}
\usage{
twosigma_custom(count, mean_form, zi_form, id, disp_covar = NULL,
  weights = rep(1, length(count)), control = glmmTMBControl())
}
\arguments{
\item{count}{Vector of non-negative integer read counts. Users should ensure that this matches matches the LHS of the formula in "mean_form."}

\item{mean_form}{Custom two-sided model formula for the (conditional) mean model. Formula is passed directly into glmmTMB with random effects specified as in the lme4 package. Users should ensure that the dependent variable matches the argument to the parameter "count."}

\item{zi_form}{Custom one-sided model formula for the zero-inflation model. Formula is passed directly into glmmTMB with random effects specified as in lme4.}

\item{id}{Vector of individual-level ID's. Used for random effect prediction.}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}.}
}
\value{
An object of class \code{glmmTMB}.
}
\description{
Fit the TWO-component SInGle cell Model-based Association method of ... with custom user-specified model formulas.
}
\section{Details}{

This function is likely only needed if users wish to include random effect terms beyond random intercepts. Users should be confident in their abilities to specify random effects using the syntax of lme4.
}

