% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigmag_anova.R
\name{twosigmag_anova}
\alias{twosigmag_anova}
\title{Gene set testing of ANOVA-style pairwise contrasts for single-cell RNA-sequencing data adjusting for inter-gene correlation using custom user-specified model formulas.}
\usage{
twosigmag_anova(count_matrix, index_test, index_ref = NULL,
  all_as_ref = FALSE, mean_form, zi_form, id, contrast_matrix, fact_name,
  rho = NULL, allow_neg_corr = FALSE, return_summary_fits = TRUE,
  disp_covar = NULL, weights = rep(1, ncol(count_matrix)),
  control = glmmTMBControl())
}
\arguments{
\item{count_matrix}{Matrix of non-negative integer read counts, with rows corresponding to genes and columns correspoding to cells. It is recommended to make the rownames the gene names for better output.}

\item{index_test}{List of indices corresponding to rows of the count matrix that are in the test set. Names of each list element (i.e. Gene Set Names) are carried forward to output if present.}

\item{index_ref}{List of indices corresponding to rows of the count matrix that are in the reference set.  If NULL, a reference set is randomly selected of the same size as the test size using genes not in the test set (if all_as_ref=FALSE) or using all other genes (if all_as_ref=TRUE). See \code{all_as_ref}. Must be either NULL or a list with the same length as index_test.}

\item{all_as_ref}{Should all genes not in the test set be used as the reference? If FALSE, a random subset is taken of size equal to the test size.}

\item{mean_form}{Custom two-sided model formula for the (conditional) mean model under the alternative. Formula is passed directly into glmmTMB with random effects specified as in the lme4 package. Users should ensure that the LHS of the formula begins with "count."}

\item{zi_form}{Custom one-sided model formula for the zero-inflation model under the alternative. Formula is passed directly into glmmTMB with random effects specified as in lme4.}

\item{id}{Vector of individual-level (sample-level) ID's. Used to estimate inter-gene correlation and for random effect prediction (if present) and required.}

\item{contrast_matrix}{Contrast matrix to test using the glht and mcp functions. Rows correspond to different contrasts, and columns refer to the ordered levels of \code{fact_name}. Rownames are carried forward to output if present.}

\item{fact_name}{Factor variable to test using \code{contrast_matrix}.}

\item{rho}{Inter-gene correlation value. If NULL (default), estimated using TWO-SIGMA model residuals.}

\item{allow_neg_corr}{Should negative correlation values be allowed? If FALSE, correlation is set to zero (leads to conservative inference).}

\item{return_summary_fits}{Should complete model fits be returned?}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model. Defaults to NULL for constant dispersion.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}.  See \code{?glmmTMBControl}.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{gene_summary_fits: }}{Summary.glmmTMB objects for each gene from the alternative model (if return_summary_fits=TRUE)}
\item{\code{z_stats_gene_level_all: }}{Matrix of z statistics testing the contrast_matrix specified in \code{contrast_matrix}. Row order matches the order of inputted test sets, and column order matches the order of \code{contrast_matrix}.}
\item{\code{p.vals_gene_level: }}{Matrix of p-values testing the contrast_matrix specified in \code{contrast_matrix}. Row order matches the order of inputted test sets, and column order matches the order of \code{contrast_matrix}.}
\item{\code{set_p.val: }}{Matrix of unadjusted set-level p-values. Row order matches the order of inputted test sets, and column order matches the order of \code{contrast_matrix}.}
\item{\code{set_p.val_ttest: }}{Matrix of unadjusted set-level p-values using the t-test. Row order matches the order of inputted test sets, and column order matches the order of \code{contrast_matrix}.}
\item{\code{estimates_gene_level: }}{Matrix of gene-level contrast_matrix estimates. Row order matches the order of inputted test sets, and column order matches the order of \code{contrast_matrix}}
\item{\code{direction: }}{Reports whether the test set tends to be Up or Down Regulated based on the covariate specified in the mean_covar_logFC argument.}
\item{\code{corr: }}{Vector of estimated inter-gene correlations for each test set. Order matches the order of inputted test sets.}
\item{\code{test_sets: }}{Vector of numeric indices corresponding to genes in each test set.}
\item{\code{ref_sets: }}{Vector of numeric indices corresponding to the genes in each reference set.}
}
}
\description{
Gene set testing of ANOVA-style pairwise contrasts for single-cell RNA-sequencing data adjusting for inter-gene correlation using custom user-specified model formulas.
}
\section{Details}{
  If adhoc=TRUE, any input in mean_re and zi_re will be ignored.
}

