% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosigmag_custom.R
\name{twosigmag_custom}
\alias{twosigmag_custom}
\title{Gene set testing adjusting for inter-gene correlation using the TWO-SIGMA model of ... with custom user-specified model formulas.}
\usage{
twosigmag_custom(count_matrix, index_test, index_ref = NULL,
  all_as_ref = FALSE, mean_form_alt, zi_form_alt, mean_form_null,
  zi_form_null, id, lr.df, rho = NULL, allow_neg_corr = FALSE,
  disp_covar = NULL, return_fits = FALSE, weights = rep(1,
  length(count_matrix[1, ])), control = glmmTMBControl())
}
\arguments{
\item{count_matrix}{Matrix of non-negative integer read counts. If specifying custom formula(s) via the arguments mean_form and zi_form the expression in mean_form will supersede.}

\item{index_test}{Index corresponding to rows of the count matrix that are in the test set. Either a list with each element being a different set or a numeric vector to test only a single set.}

\item{index_ref}{Index corresponding to rows of the count matrix that are in the reference set.  If NULL, a reference set is randomly selected of the same size as test size using genes not in the test set or using all other genes. See all_as_ref.}

\item{all_as_ref}{Should all genes not in the test set be used as the reference? If FALSE, a random subset is taken of size equal to the test size.}

\item{mean_form_alt}{Custom two-sided model formula for the (conditional) mean model under the null. Formula is passed directly into glmmTMB with random effects specified as in the lme4 package. Users should ensure that the dependent variable matches the argument to the parameter "count."}

\item{zi_form_alt}{Custom one-sided model formula for the zero-inflation model under the alternative. Formula is passed directly into glmmTMB with random effects specified as in lme4.}

\item{mean_form_null}{Custom two-sided model formula for the (conditional) mean model under the null. Syntax is as in \code{mean_form_alt}.}

\item{zi_form_null}{Custom one-sided model formula for the zero-inflation model under the null. Syntax is as in \code{zi_form_alt}.}

\item{id}{Vector of individual-level ID's. Used for random effect prediction and the adhoc method but required regardless.}

\item{lr.df}{degrees of freedom for the asymptotic chi-square approximation to the liklihood ratio statistic.}

\item{rho}{Inter-gene correlation value. If NULL (default), estimated using TWO-SIGMA model residuals.}

\item{allow_neg_corr}{Should negative correlation values be allowed? If FALSE, correlation is set to zero (leads to conservative inference).}

\item{disp_covar}{Covariates for a log-linear model for the dispersion. Either a matrix of covariates or = 1 to indicate an intercept only model. Random effect terms are not permitted in the dispersion model. Defaults to NULL for constant dispersion.}

\item{return_fits}{Should complete model fits be returned? Use cautiously because fit objects can be very large.}

\item{weights}{weights, as in glm. Defaults to 1 for all observations and no scaling or centering of weights is performed.}

\item{control}{Control parameters for optimization in \code{glmmTMB}.}

\item{LR_stats_gene_level_all:}{Gives all gene-level likelihood ratio statistics.  Order matches the order of the inputted count matrix}

\item{p.vals_gene_level:}{Gives p-values associated with `LR_stats_gene_level_all`.}

\item{set_p.val:}{Vector of unadjusted set-level p-values. Order matches the order of inputted test sets.}

\item{corr:}{Vector of estimated inter-gene correlations for each test set. Order matches the order of inputted test sets.}

\item{test_sets:}{Vector of numeric indices corresponding to genes in each test set.}

\item{ref_sets:}{Vector of numeric indices corresponding to the genes in each reference set.}

\item{gene_level_fits:}{List of gene-level fits, returned only if `return_fits` is TRUE.}
}
\value{
An object of class \code{glmmTMB}.
}
\description{
Gene set testing adjusting for inter-gene correlation using the TWO-SIGMA model of ... with custom user-specified model formulas.
}
\section{Details}{
  If adhoc=TRUE, any input in mean_re and zi_re will be ignored.
}

